Run the test suite in C# instead of Java
This commit updates the test suite to only generate C# code. The compilation
and execution of the test code is left for the various Antlr4.Runtime.Test
projects. The strategy results in a 20x improvement for generating and
executing the test suite, and over a 100x improvement for rerunning tests
without generating them again.

This work is heavily based on the approach used by @BurtHarris in
tunnelvisionlabs/antlr4ts#244.