We hit an issue where C# generator tool would exit with no error message and no files generated.
After troubleshooting I root-caused the issue to .NET command line parsing rules, where special characters like double-quotes and backslashes have to be escaped under some circumstances.

There already was a code to handle some of these cases, but it didn't cover the following:
1) When there are multiple backslashes preceeding the double-quote, we need to escape all of them, not only the last one.
2) When there is a backslash at the end of the argument, we need to escape it, otherwise it will escape the double-quote appended by the code itself, causing attribute value to run out when parsed by .NET infrastructure.

This change is improving the logic to account for these two extra cases. I also added unit tests to demonstrate the behavior. In order to do so, I had to move the JoinArguments methd to a public helper class, so that it can be accessed from unit tests, as I'm not a big fan of InternalsVisibleTo.